requirements.txt
python-telegram-bot
supabase
python-dotenv
Flask